Refactor: Split the Docker Hub release module
release_docker_hub.py was 976 lines holding the whole Nexus3 to Docker
Hub workflow: configuration, HTTP helpers, three tag classes, the
per-repository copy machinery, catalogue collection and the reporting
functions.

Most of the module communicates through module-level state that the
test suite assigns directly, which is what made this awkward to split.
Anything moved out of the root module would read a stale copy of a
setting rebound after import.

Resolve that by giving the settings their own module. Readers now
access them as attributes of that module, so they observe the value
current at call time rather than the value bound when their own module
was imported. The names lose their leading underscore because they are
no longer private to a single file.

The remaining split follows what depends on what:

  settings.py  nexus3 URLs, the version regexp and their setters
  util.py      URL and image-id string helpers, repo file lookups
  tags.py      TagClass, NexusTagClass, DockerTagClass
  project.py   ProjectClass and the retry table it drives
  __init__.py  catalogue collection, reporting and the entry point

The mutable NexusCatalog and projects lists stay in the root module,
where callers and tests already reach them. The root module re-exports
the moved names, so lftools_uv.nexus.release_docker_hub keeps the same
public surface; __all__ now records what that surface is.

Two test assertions move from rdh._version_regexp to
rdh.settings.version_regexp, following the value.
