+++ tests/test_release_docker_hub.py
@@ -328,14 +328,14 @@ class TestTagsRegExpClass:
         test_regexp_from_file = os.path.join(str(datafiles), "releasedockerhub_good_regexp")
         rdh.initialize(org, test_regexp_from_file)
         assert rdh.validate_regexp()
-        assert rdh._version_regexp == r"^\d+.\d+"
+        assert rdh.settings.version_regexp == r"^\d+.\d+"
 
     def test_tag_class_manual_version_regexp_str_from_file_invalid(self, datafiles):
         org = "onap"
         test_regexp_from_file = os.path.join(str(datafiles), "releasedockerhub_bad_regexp")
         rdh.initialize(org, test_regexp_from_file)
         assert not rdh.validate_regexp()
-        assert rdh._version_regexp == "["
+        assert rdh.settings.version_regexp == "["
 
 
 @pytest.mark.datafiles(
